chore(cast): reduce session revoke diff
